# tsdb: more efficient sorting of postings read from WAL at startup