@bboreham 2022-04-13
It seems to take several hours to kick in, which is interesting if it's caused by a change that only runs during startup.

![image](https://user-images.githubusercontent.com/8125524/163146978-e51ceec3-9350-4565-8000-5e0191a9c01c.png)